# docs/swagger.yaml
basePath: /v1
definitions:
  models.AddEducationParameter:
    properties:
      education:
        items:
          $ref: '#/definitions/models.EducationParameters'
        type: array
    type: object
  models.AddExperienceParameter:
    properties:
      experience:
        items:
          $ref: '#/definitions/models.ExperienceParameters'
        type: array
    type: object
...
  models.EducationParameters:
    properties:
      endYear:
        type: integer
      name:
        type: string
      startYear:
        type: integer
    type: object
  models.EducationReturnValue:
    properties:
      endYear:
        type: integer
      id:
        type: integer
      name:
        type: string
      startYear:
        type: integer
    type: object
  models.ExperienceParameters:
    properties:
      endYear:
        type: integer
      place:
        type: string
      position:
        type: string
      startYear:
        type: integer
    type: object
  models.ExperienceReturnValue:
    properties:
      endYear:
        type: integer
      id:
        type: integer
      place:
        type: string
      position:
        type: string
      startYear:
        type: integer
    type: object
...
  models.PortfolioEditParameter:
    properties:
      description:
        type: string
    type: object
...
  models.ResponseOKGetUserEducation:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/models.EducationReturnValue'
        type: array
      message:
        type: string
    type: object
  models.ResponseOKGetUserExperience:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/models.ExperienceReturnValue'
        type: array
      message:
        type: string
    type: object
...
  models.ResponseWithNoBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  models.UpdateSkills:
    properties:
      skills:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    name: API Support
  description: Swagger API for Golang Project PMS + Freelancer.
  license: {}
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /addEducation:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Data Format to add education
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/models.AddEducationParameter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Add User Education
      tags:
      - Education
  /addExperience:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Data Format to add experience
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/models.AddExperienceParameter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Adding User Experience
      tags:
      - Experience
  /addPortfolio:
    post:
      consumes:
      - application/json
      - multipart/form-data
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Upload File
        in: formData
        name: file
        required: true
        type: file
      - description: Description of the File
        in: formData
        name: description
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Adding User Portfolio
      tags:
      - Portfolio
...
  /allSkills:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Getting all list of skills
      tags:
      - Skills
...
  /deletePortfolio/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Deleting User Portfolio
      tags:
      - Portfolio
  /editPortfolio/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: New Portfolio Description
        in: body
        name: Description
        required: true
        schema:
          $ref: '#/definitions/models.PortfolioEditParameter'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Updating User Portfolio
      tags:
      - Portfolio
  /editSkills:
    put:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      - description: New Skills ID List
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSkills'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Updating User Skills
      tags:
      - Skills
...
  /userEducation:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOKGetUserEducation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: User Education
      tags:
      - Education
  /userExperience:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOKGetUserExperience'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: User Experience
      tags:
      - Experience
...
swagger: "2.0"